requirements.txt
appdirs==1.4.4
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
cmake==3.15.3
cycler==0.10.0
decorator==4.4.1
distlib==0.3.0
Django==3.0.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
dlib==19.18.0
et-xmlfile==1.0.1
filelock==3.0.12
googletrans==2.4.0
idna==2.8
imageio==2.6.1
imutils==0.5.3
jdcal==1.4.1
joblib==0.14.1
kiwisolver==1.1.0
networkx==2.4
protobuf==3.11.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
urllib3==1.25.7
virtualenv==20.0.21
